rename logic data modules
