+++ logic_area.py
