+++ logic_locations.py
